Use AspectJ Maven instead of jcabi weaving plugin for binary weaving
AJ Maven is the recommended plugin for such things, not an exotic plugin
providing no added value here.

We also do not need @Configuration class AspectConfig, because a
native AspectJ aspect is not a Spring-managed component.

The console log running the test should contain something like this:

service 0.0839
Hello World!